PRO-2672 Move use :model behavioral specs to spec_rails; test the no-AR guard
Requiring active_record is process-wide, so keeping the model logic specs in
spec/ leaked AR into the whole non-Rails suite and made the "raises without
ActiveRecord" path untestable. Relocate all behavioral coverage to the Rails
suite (against the real User/Profile models) and leave spec/ deliberately
AR-free — where it now asserts use :model raises NotImplementedError, with a
defined?(ActiveRecord) canary guarding that no other spec reintroduces an AR
require.

General convention going forward: anything needing ActiveRecord lives in
spec_rails/, so spec/ stays a genuine non-Rails environment.
